feat: add native logic
