+++ package.json
@@ -18,7 +18,9 @@
     "test:unit": "vitest --coverage --run",
     "test.local.init": "rimraf localBaseline && wdio ./tests/configs/wdio.local.init.conf.ts",
     "test.local.desktop": "wdio tests/configs/wdio.local.desktop.conf.ts",
+    "test.local.emus.app": "wdio tests/configs/wdio.local.android.emus.app.conf.ts",
     "test.local.emus.web": "wdio tests/configs/wdio.local.android.emus.web.conf.ts",
+    "test.local.sims.app": "wdio tests/configs/wdio.local.ios.sims.app.conf.ts",
     "test.local.sims.web": "wdio tests/configs/wdio.local.ios.sims.web.conf.ts",
     "test.local.desktop.multi": "wdio tests/configs/wdio.local.desktop.multi.conf.ts",
     "test.unit.coverage": "vitest --coverage",
