+++ packages/webdriver-image-comparison/src/commands/checkAppElement.ts
@@ -1,0 +1,10 @@
+import type { ImageCompareResult } from '../methods/images.interfaces.js'
+
+/**
+ * Compare  an image of the element
+ */
+export default async function checkAppElement(
+): Promise<ImageCompareResult | number> {
+    console.log('\n\nNATIVE CHECK ELEMENT SCREENSHOT\n\n')
+    return 0
+}
