+++ packages/webdriver-image-comparison/src/commands/checkAppScreen.ts
@@ -1,0 +1,11 @@
+import type {  ImageCompareResult } from '../methods/images.interfaces.js'
+
+/**
+ * Compare an image of the viewport of the screen
+ */
+export default async function checkAppScreen(
+): Promise<ImageCompareResult | number> {
+    console.log('\n\nNATIVE CHECK SCREEN SCREENSHOT\n\n')
+
+    return 0
+}
